md endpoints + accept + discoverability
