+++ src/pages/ccip/[...id].astro
@@ -1,56 +1,78 @@
 ---
+export const prerender = false
+
 import DocsLayout from "~/layouts/DocsLayout.astro"
 import { getCollection, render } from "astro:content"
+import { buildMarkdownDocumentFromPath, markdownHeaders } from "@lib/markdown/buildMarkdownResponse.js"
 
-export async function getStaticPaths() {
-  const ccipEntries = await getCollection("ccip")
-  const pathMap = new Map()
+function mapCcipRouteId(entryId: string): string | null {
+  const urlPath = entryId.replace(/\.(md|mdx)$/, "")
 
-  // Process all entries to build URL path mapping
-  for (const entry of ccipEntries) {
-    const urlPath = entry.id.replace(/\.(md|mdx)$/, "")
+  // Skip directory and tutorial entries (handled by dedicated routes)
+  if (urlPath.startsWith("directory/") || urlPath.startsWith("tutorials/")) {
+    return null
+  }
 
-    // Skip directory and tutorial entries (handled by dedicated routes)
-    if (urlPath.startsWith("directory/") || urlPath.startsWith("tutorials/")) {
-      continue
-    }
+  // Handle API reference version format conversion (v150 -> v1.5.0)
+  if (urlPath.includes("/api-reference/") && urlPath.includes("/v")) {
+    const parts = urlPath.split("/")
+    const versionIndex = parts.findIndex((part) => part.startsWith("v") && /^v\d+$/.test(part))
 
-    // Handle API reference version format conversion (if needed)
-    if (urlPath.includes("/api-reference/") && urlPath.includes("/v")) {
-      const parts = urlPath.split("/")
-      const versionIndex = parts.findIndex((part) => part.startsWith("v") && /^v\d+$/.test(part))
-
-      if (versionIndex !== -1) {
-        const versionPart = parts[versionIndex]
-        // Convert version format if needed (v150 → v1.5.0)
-        if (versionPart.length === 4) {
-          const formattedVersion = `v${versionPart[1]}.${versionPart[2]}.${versionPart[3]}`
-          parts[versionIndex] = formattedVersion
-          pathMap.set(parts.join("/"), entry)
-        } else {
-          pathMap.set(urlPath, entry)
-        }
-      } else {
-        pathMap.set(urlPath, entry)
+    if (versionIndex !== -1) {
+      const versionPart = parts[versionIndex]
+      if (versionPart.length === 4) {
+        const formattedVersion = `v${versionPart[1]}.${versionPart[2]}.${versionPart[3]}`
+        parts[versionIndex] = formattedVersion
+        return parts.join("/")
       }
-    } else {
-      pathMap.set(urlPath, entry)
     }
   }
 
-  // Generate static paths
-  return Array.from(pathMap.entries()).map(([urlPath, entry]) => ({
-    params: { id: urlPath },
-    props: { entry },
-  }))
+  return urlPath
 }
 
-interface Props {
-  entry: Awaited<ReturnType<typeof getCollection<"ccip">>>[number]
+const requestedId = Astro.params.id?.replace(/^\/+|\/+$/g, "") || "index"
+
+const ccipEntries = await getCollection("ccip")
+
+const entry = ccipEntries.find((candidate) => {
+  const mapped = mapCcipRouteId(candidate.id)
+  return mapped === requestedId
+})
+
+if (!entry) {
+  return new Response("Page not found.", { status: 404 })
 }
 
-const { entry } = Astro.props
+const accept = Astro.request.headers.get("accept") || ""
+const wantsMarkdown = accept
+  .toLowerCase()
+  .split(",")
+  .map((part) => part.split(";")[0]?.trim())
+  .includes("text/markdown")
+
+const canonicalPath = requestedId === "index" ? "ccip" : `ccip/${requestedId}`
+
+if (wantsMarkdown) {
+  const markdown = await buildMarkdownDocumentFromPath(canonicalPath, Astro.request)
+
+  if (!markdown) {
+    return new Response("Page not found.", { status: 404 })
+  }
+
+  return new Response(markdown, {
+    status: 200,
+    headers: {
+      ...markdownHeaders,
+      Vary: "Accept",
+    },
+  })
+}
+
 const { Content, headings } = await render(entry)
+
+Astro.response.headers.append("Link", `</${canonicalPath}.md>; rel="alternate"; type="text/markdown"`)
+Astro.response.headers.set("Vary", "Accept")
 ---
 
 <DocsLayout frontmatter={entry.data} {headings}>
