+++ src/pages/ccip/tutorials/[...id].astro
@@ -1,35 +1,60 @@
 ---
+export const prerender = false
+
 import { getCollection, render, type CollectionEntry } from "astro:content"
 import TutorialLayout from "../../../layouts/TutorialLayout.astro"
 import DocsLayout from "../../../layouts/DocsLayout.astro"
+import { buildMarkdownDocumentFromPath, markdownHeaders } from "@lib/markdown/buildMarkdownResponse.js"
 
 type Props = {
   entry: CollectionEntry<"ccip">
 }
 
-export async function getStaticPaths() {
-  const entries = await getCollection("ccip", (entry) => entry.id.startsWith("tutorials/"))
+const requestedId = Astro.params.id?.replace(/^\/+|\/+$/g, "") || "index"
 
-  const paths: { params: { id: string }; props: Props }[] = []
+const entries = await getCollection("ccip", (entry) => entry.id.startsWith("tutorials/"))
 
-  for (const entry of entries) {
-    const tutorialId = entry.id.replace(/^tutorials\//, "").replace(/\.(md|mdx)$/, "")
+const entry = entries.find((candidate) => {
+  const tutorialId = candidate.id.replace(/^tutorials\//, "").replace(/\.(md|mdx)$/, "")
+  return tutorialId === requestedId
+})
 
-    paths.push({
-      params: { id: tutorialId },
-      props: { entry },
-    })
-  }
-
-  return paths
+if (!entry) {
+  return new Response("Page not found.", { status: 404 })
 }
 
-const { entry } = Astro.props as Props
+const accept = Astro.request.headers.get("accept") || ""
+const wantsMarkdown = accept
+  .toLowerCase()
+  .split(",")
+  .map((part) => part.split(";")[0]?.trim())
+  .includes("text/markdown")
+
+const canonicalPath = requestedId === "index" ? "ccip/tutorials" : `ccip/tutorials/${requestedId}`
+
+if (wantsMarkdown) {
+  const markdown = await buildMarkdownDocumentFromPath(canonicalPath, Astro.request)
+
+  if (!markdown) {
+    return new Response("Page not found.", { status: 404 })
+  }
+
+  return new Response(markdown, {
+    status: 200,
+    headers: {
+      ...markdownHeaders,
+      Vary: "Accept",
+    },
+  })
+}
 
 const { Content, headings } = await render(entry)
 
-// Only use TutorialLayout for specific interactive tutorials
+// Preserve original layout switch
 const Layout = entry.id === "tutorials/evm/cross-chain-tokens/register-from-eoa-remix" ? TutorialLayout : DocsLayout
+
+Astro.response.headers.append("Link", `</${canonicalPath}.md>; rel="alternate"; type="text/markdown"`)
+Astro.response.headers.set("Vary", "Accept")
 ---
 
 <Layout frontmatter={entry.data} headings={headings}>
